added clear color as parameter & fixed animated sprite

+++ include/MOGL/AnimatedSprite.hpp
@@ -17,7 +17,7 @@ class AnimatedSprite : public Sprite
 {
 public:
     enum class Status { PLAYING, PAUSED, STOPPED };
-    AnimatedSprite(int width, int height, const SpriteAnimation* animation);
+    AnimatedSprite(double width, double height, const SpriteAnimation* animation);
     const SpriteAnimation* spriteAnimation() const;
     void setSpriteAnimation(const SpriteAnimation* animation);
